ai-chatbackend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_community
langchain-core
langchain-groq
groq
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pyjwt
python-jose